made packeges to process data from cli
