+++ src/config
@@ -1,6 +1,6 @@
 [DEFAULT]
-MODEL = prajjwal1/bert-tiny
-BATCH_SIZE = 8
-num_warmup_steps = 0
-num_train_epochs = 3
-output_dir = ./models
+    MODEL = prajjwal1/bert-tiny
+    BATCH_SIZE = 8
+    num_warmup_steps = 0
+    num_train_epochs = 3
+    output_dir = ./models
