+++ src/data/make_dataset.py
@@ -1,18 +1,14 @@
 # -*- coding: utf-8 -*-
-import configparser
-import numpy as np
-from datasets import load_dataset
-from torch.utils.data import DataLoader
+import fire
+from datasets import load_from_disk
 from transformers import AutoTokenizer
 from transformers import DataCollatorForTokenClassification
+import configparser
 
-raw_datasets = load_dataset("conll2003")
+# model_checkpoint = config["DEFAULT"]["MODEL"]
 config = configparser.ConfigParser()
 config.read('../config')
-
-model_checkpoint = config["DEFAULT"]["MODEL"]
-ner_feature = raw_datasets["train"].features["ner_tags"]
-label_names = ner_feature.feature.names
+model_checkpoint = "prajjwal1/bert-tiny"
 
 tokenizer = AutoTokenizer.from_pretrained(model_checkpoint)
 data_collator = DataCollatorForTokenClassification(tokenizer=tokenizer)
@@ -55,29 +51,15 @@ def tokenize_and_align_labels(examples):
     return tokenized_inputs
 
 
-def make_dataloader(batch_size, data_type="train"):
+def make_dataset(input_path, output_path):
+    raw_datasets = load_from_disk(input_path)
     tokenized_datasets = raw_datasets.map(
         tokenize_and_align_labels,
         batched=True,
-        remove_columns=raw_datasets[data_type].column_names,
+        remove_columns=raw_datasets["train"].column_names,
     )
-    dataloader = DataLoader(
-        tokenized_datasets[data_type],
-        shuffle=True,
-        collate_fn=data_collator,
-        batch_size=batch_size,
-    )
-    return dataloader
+    tokenized_datasets.save_to_disk(output_path)
 
 
-def postprocess(predictions: np.array, labels: np.array):
-    predictions = predictions.detach().cpu().clone().numpy()
-    labels = labels.detach().cpu().clone().numpy()
-
-    # Remove ignored index (special tokens) and convert to labels
-    true_labels = [[label_names[lbl] for lbl in label if lbl != -100] for label in labels]
-    true_predictions = [
-        [label_names[p] for (p, lbl) in zip(prediction, label) if lbl != -100]
-        for prediction, label in zip(predictions, labels)
-    ]
-    return true_labels, true_predictions
+if __name__ == '__main__':
+    fire.Fire(make_dataset)
